[2022 Day 11 Part 2] Parse the monkeys from the input files.
For expedience, my original solve was hardcoded into the source. I achieved it by pasting the input (which was almost code) into Python and then using several vim sed operations to massage it into actual Python.

This change replaces the use of vim sed with actual regular expressions. Importantly, it uses eval() for the worry operation and divisibility tests.

Due to the use of eval() it actually takes a performance hit, but it's in the greater spirit of my other solves to parse the input.

To regain the performance I'll have to actually parse the operations and eliminate the use of eval(), but I doubt I'll get to it.